Version 0.3: spread code and pep8
