+++ __init__.py
@@ -1,5 +1,5 @@
 """
-Copyright (C) 2016 Marco Pasi <mf.pasi@gmail.com> 
+Copyright (C) 2016 Marco Pasi <mf.pasi@gmail.com>
 
 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -9,10 +9,11 @@ the Free Software Foundation; either version 2 of the License, or
 
 from app import App
 from metadata import Metadata
-from tool import Tool, Workflow
+from tool import Tool
+from workflow import Workflow
 import apps
 import examples
 
-__author__  = 'Marco Pasi, Javier Conejero'
-__version__ = 'v0.2'
+__author__ = 'Marco Pasi, Javier Conejero'
+__version__ = '0.3'
 __license__ = 'GPL v2'
