+++ apps/workflowapp.py
@@ -1,8 +1,9 @@
-#------------------------------------------------------------------------------
+# ------------------------------------------------------------------------------
 # Workflow App
-#------------------------------------------------------------------------------
+# ------------------------------------------------------------------------------
 from . import PyCOMPSsApp, LocalApp
 from .. import Workflow
+
 
 class WorkflowApp(PyCOMPSsApp, LocalApp):
     """
